requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
#celery==4.4.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-beat==2.2.1
django-cors-headers==3.10.0
django-jet==1.0.8
django-timezone-field==4.2.1
djangorestframework==3.12.4
importlib-metadata==4.8.1
kombu==4.6.11
#Pillow==8.3.2
#pkg-resources==0.0.0
prompt-toolkit==3.0.20
#psycopg2==2.9.3
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
vine==1.3.0
wcwidth==0.2.5
zipp==3.6.